+++ resources/views/bible/index.blade.php
@@ -2,15 +2,27 @@
 
 @section('content')
     <div class="container">
-        <div class="app">
+        <div id="app">
             {{-- <router-link to="/read/example">Go to Example</router-link> --}}
 
             {{-- <router-view></router-view> --}}
             <index-component :books="books" v-on:select-page="selectPage"></index-component>
+
+            <?php
+                $loggedIn = Auth::check() ? 'true' : 'false';
+                
+            ?>
             <page-component 
                 :page="selectedPage" 
-                :chapter-text="pageText"   
+                :chapter-text="pageText"
+                logged-in="{{ $loggedIn }}"
                 v-on:leaf-page="leafPage"></page-component>
         </div>
     </div>
 @endsection
+
+{{-- Added for testing. --}}
+<?php
+ $user = auth()->user();
+ var_dump($user);
+?>